Rename and add missing RoomSettings Events
