+++ Communication/Packets/Incoming/RoomSettings/GetFlatControllersEvent.cs
@@ -1,0 +1,20 @@
+﻿using Plus.Communication.Packets.Outgoing.Rooms.Settings;
+using Plus.HabboHotel.GameClients;
+
+namespace Plus.Communication.Packets.Incoming.RoomSettings;
+
+internal class GetFlatControllersEvent : IPacketEvent
+{
+    public Task Parse(GameClient session, IIncomingPacket packet)
+    {
+        if (!session.GetHabbo().InRoom)
+            return Task.CompletedTask;
+        var instance = session.GetHabbo().CurrentRoom;
+        if (instance == null)
+            return Task.CompletedTask;
+        if (!instance.CheckRights(session))
+            return Task.CompletedTask;
+        session.Send(new RoomRightsListComposer(instance));
+        return Task.CompletedTask;
+    }
+}
