--- Communication/Packets/Incoming/Rooms/Settings/GetRoomRightsEvent.cs
@@ -1,20 +1,0 @@
-﻿using Plus.Communication.Packets.Outgoing.Rooms.Settings;
-using Plus.HabboHotel.GameClients;
-
-namespace Plus.Communication.Packets.Incoming.Rooms.Settings;
-
-internal class GetRoomRightsEvent : IPacketEvent
-{
-    public Task Parse(GameClient session, IIncomingPacket packet)
-    {
-        if (!session.GetHabbo().InRoom)
-            return Task.CompletedTask;
-        var instance = session.GetHabbo().CurrentRoom;
-        if (instance == null)
-            return Task.CompletedTask;
-        if (!instance.CheckRights(session))
-            return Task.CompletedTask;
-        session.Send(new RoomRightsListComposer(instance));
-        return Task.CompletedTask;
-    }
-}
